Added stars for reviews and made sure the form works
